Refactor  cd prefix helper
