+++ src/vs/workbench/contrib/terminalContrib/chatAgentTools/browser/runInTerminalHelpers.ts
@@ -300,3 +300,35 @@ export function dedupeRules(rules: ICommandApprovalResultWithReason[]): ICommand
 		return array.findIndex(r => isAutoApproveRule(r.rule) && r.rule.sourceText === sourceText) === index;
 	});
 }
+
+export interface IExtractedCdPrefix {
+	/** The directory path that was extracted from the cd command */
+	directory: string;
+	/** The command to run after the cd */
+	command: string;
+}
+
+/**
+ * Extracts a cd prefix from a command line, returning the directory and remaining command.
+ * Does not check if the directory matches the current cwd - just extracts the pattern.
+ */
+export function extractCdPrefix(commandLine: string, shell: string, os: OperatingSystem): IExtractedCdPrefix | undefined {
+	const isPwsh = isPowerShell(shell, os);
+
+	const cdPrefixMatch = commandLine.match(
+		isPwsh
+			? /^(?:cd(?: \/d)?|Set-Location(?: -Path)?) (?<dir>[^\s]+) ?(?:&&|;)\s+(?<suffix>.+)$/i
+			: /^cd (?<dir>[^\s]+) &&\s+(?<suffix>.+)$/
+	);
+	const cdDir = cdPrefixMatch?.groups?.dir;
+	const cdSuffix = cdPrefixMatch?.groups?.suffix;
+	if (cdDir && cdSuffix) {
+		// Remove any surrounding quotes
+		let cdDirPath = cdDir;
+		if (cdDirPath.startsWith('"') && cdDirPath.endsWith('"')) {
+			cdDirPath = cdDirPath.slice(1, -1);
+		}
+		return { directory: cdDirPath, command: cdSuffix };
+	}
+	return undefined;
+}
